# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PAY API",
...
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user-virtual-account/get-list": {
            "post": {
                "tags": [
                    "Virtual Account"
                ],
                "summary": "Danh sách VA",
                "operationId": "19e3100e4768a1a5e07bdd3e0082998f",
                "parameters": [
                    {
                        "name": "api-token",
                        "in": "header",
                        "description": "token allow create service",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sign": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "sign": "signiture data"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result_false": {
                                        "summary": "Result error.",
                                        "value": {
                                            "error_code": 401,
                                            "system_time": "2023-11-24 11:15:59",
                                            "message": "access_token hết hạn hoặc không được phép sử dụng",
                                            "errors": [
                                                "access_token hết hạn hoặc không được phép sử dụng"
                                            ],
                                            "data": null,
                                            "tranid_tracking": "bc47d534b97f7cdb4c2d5d229376915c"
                                        }
                                    },
                                    "result_success": {
                                        "summary": "Result success.",
                                        "value": {
                                            "error_code": 200,
                                            "message": "success",
                                            "system_time": "2023-11-24 11:15:59",
                                            "errors": [],
                                            "data": []
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}